Add test_existing_file to live tests
